feat: регистрация и часть с юзером готова
