+++ src/components/Movies/Movies.jsx
@@ -4,76 +4,76 @@ import Preloader from '../Preloader/Preloader.jsx';
 import MoviesCardList from '../MoviesCardList/MoviesCardList.jsx';
 import './Movies.css';
 
-function Movies () {
+function Movies ({loggedIn,  cards}) {
     const load = false;
-    const moviesCard = [
-        {
-            id: 1,
-            title: '33 слова о дизайне',
-            time: '1ч 47м',
-            image:'https://images.unsplash.com/photo-1666812272845-c5dcaae45453?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=387&q=80',
-        },
-        {
-            id: 2,
-            title: '33 слова о дизайне',
-            time: '1ч 47м',
-            image:'https://images.unsplash.com/photo-1666668321985-105042873ddc?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=387&q=80',
-        },
-        {
-            id: 3,
-            title: '33 слова о дизайне',
-            time: '1ч 47м',
-            image:'https://images.unsplash.com/photo-1666834744271-af5ef132a179?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=870&q=80',
-        },
-        {
-            id: 4,
-            title: '33 слова о дизайне',
-            time: '1ч 47м',
-            image:'https://images.unsplash.com/photo-1666625628272-a1071f6f7173?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1032&q=80',
-        },
-        {
-            id: 5,
-            title: '33 слова о дизайне',
-            time: '1ч 47м',
-            image:'https://images.unsplash.com/photo-1666625628272-a1071f6f7173?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1032&q=80',
-        },
-        {
-            id: 6,
-            title: '33 слова о дизайне',
-            time: '1ч 47м',
-            image:'https://images.unsplash.com/photo-1666625628272-a1071f6f7173?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1032&q=80',
-        },
-        {
-            id: 7,
-            title: '33 слова о дизайне',
-            time: '1ч 47м',
-            image:'https://images.unsplash.com/photo-1666625628272-a1071f6f7173?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1032&q=80',
-        },
-        {
-            id: 8,
-            title: '33 слова о дизайне',
-            time: '1ч 47м',
-            image:'https://images.unsplash.com/photo-1666625628272-a1071f6f7173?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1032&q=80',
-        },
-        {
-            id: 9,
-            title: '33 слова о дизайне',
-            time: '1ч 47м',
-            image:'https://images.unsplash.com/photo-1666625628272-a1071f6f7173?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1032&q=80',
-        },
-        {
-            id: 10,
-            title: '33 слова о дизайне',
-            time: '1ч 47м',
-            image:'https://images.unsplash.com/photo-1666625628272-a1071f6f7173?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1032&q=80',
-        },
-        {
-            id: 11,
-            title: '33 слова о дизайне',
-            time: '1ч 47м',
-            image:'https://images.unsplash.com/photo-1666625628272-a1071f6f7173?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1032&q=80',
-        },
-    ]
+    // const moviesCard = [
+    //     {
+    //         id: 1,
+    //         title: '33 слова о дизайне',
+    //         time: '1ч 47м',
+    //         image:'https://images.unsplash.com/photo-1666812272845-c5dcaae45453?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=387&q=80',
+    //     },
+    //     {
+    //         id: 2,
+    //         title: '33 слова о дизайне',
+    //         time: '1ч 47м',
+    //         image:'https://images.unsplash.com/photo-1666668321985-105042873ddc?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=387&q=80',
+    //     },
+    //     {
+    //         id: 3,
+    //         title: '33 слова о дизайне',
+    //         time: '1ч 47м',
+    //         image:'https://images.unsplash.com/photo-1666834744271-af5ef132a179?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=870&q=80',
+    //     },
+    //     {
+    //         id: 4,
+    //         title: '33 слова о дизайне',
+    //         time: '1ч 47м',
+    //         image:'https://images.unsplash.com/photo-1666625628272-a1071f6f7173?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1032&q=80',
+    //     },
+    //     {
+    //         id: 5,
+    //         title: '33 слова о дизайне',
+    //         time: '1ч 47м',
+    //         image:'https://images.unsplash.com/photo-1666625628272-a1071f6f7173?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1032&q=80',
+    //     },
+    //     {
+    //         id: 6,
+    //         title: '33 слова о дизайне',
+    //         time: '1ч 47м',
+    //         image:'https://images.unsplash.com/photo-1666625628272-a1071f6f7173?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1032&q=80',
+    //     },
+    //     {
+    //         id: 7,
+    //         title: '33 слова о дизайне',
+    //         time: '1ч 47м',
+    //         image:'https://images.unsplash.com/photo-1666625628272-a1071f6f7173?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1032&q=80',
+    //     },
+    //     {
+    //         id: 8,
+    //         title: '33 слова о дизайне',
+    //         time: '1ч 47м',
+    //         image:'https://images.unsplash.com/photo-1666625628272-a1071f6f7173?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1032&q=80',
+    //     },
+    //     {
+    //         id: 9,
+    //         title: '33 слова о дизайне',
+    //         time: '1ч 47м',
+    //         image:'https://images.unsplash.com/photo-1666625628272-a1071f6f7173?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1032&q=80',
+    //     },
+    //     {
+    //         id: 10,
+    //         title: '33 слова о дизайне',
+    //         time: '1ч 47м',
+    //         image:'https://images.unsplash.com/photo-1666625628272-a1071f6f7173?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1032&q=80',
+    //     },
+    //     {
+    //         id: 11,
+    //         title: '33 слова о дизайне',
+    //         time: '1ч 47м',
+    //         image:'https://images.unsplash.com/photo-1666625628272-a1071f6f7173?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1032&q=80',
+    //     },
+    // ]
 
     return(
         <section className="movies">
@@ -81,7 +81,7 @@ function Movies () {
             { load ? 
             <Preloader /> : 
             <MoviesCardList
-            moviesCard={moviesCard}
+            cards={cards}
             moreButtonVisibility = {true} />
 }
         </section>
