+++ src/components/MoviesCard/MoviesCard.jsx
@@ -7,7 +7,7 @@ function MoviesCard ({card}) {
         <li className='movies-card'>
             <div className='movies-card__data'>
                 <ul className='movies-card__data-input'>
-                <li className='movies-card__title'>{card.title}</li>
+                <li className='movies-card__title'>{card.nameRU}</li>
                 <li className='movies-card__time'>{card.time}</li>
                 </ul>
                 <button className='movies-card__flag' type='button'>
